Fix fixmes (#63)
* use only namebeg/nameend from the counters struct

* turned timers into a local variable

* made cdata a local variable

* made delays a local variable

* made msg_data variable local

* made regexps a local variable

* made fqdn a local variable

* fixed LGTM warning about int comparing

* fixed two LGTM warnings about ctime()

* removed global usern variable

+++ src/auth.c
@@ -32,7 +32,7 @@ unsigned int nonce_count;
 
 /* converts a hash into hex output
    taken from the RFC 2617 */
-void cvt_hex(unsigned char *_b, unsigned char *_h, int length)
+void cvt_hex(unsigned char *_b, unsigned char *_h, unsigned short length)
 {
         unsigned short i;
